Add CustomFields class with methods for operating on custom fields
Left out create and delete for now, because I can't get create to work with
the API; Pardot returns err code 10000 "Please provide a valid type for
this field" despite passing apparently valid values for type.

Updated request interceptor to pass booleans as 1 and 0, because the Pardot
API stores false as true, presumably because it's treating the value as a
string rather than a boolean.

DRYed up tests a bit.